Testing

+++ file_rename.py
@@ -1,7 +1,18 @@
 #!python3
 import re, glob, os
 
+def main():
+    print(os.path.basename(__file__), "running...")
+
+    renamer("*.jpg-large", r"^(.*)\.jpg-large$", r"\1.jpg")
+    renamer("*.unsafe", r"^(.*)\.unsafe$", r"\1")
+    # renamer("*.doc", r"^(.*)\.doc$", r"new(\1).doc") #forward
+    # renamer("*.doc", r"^new\((.*)\)\.doc", r"\1.doc") #reverse
+    print(os.path.basename(__file__), "complete...")
+
 def renamer(files, pattern, replacement):
+    isRemove = True
+
     for pathname in glob.glob(files):
         basename = os.path.basename(pathname)
         new_filename = re.sub(pattern, replacement, basename)
@@ -11,16 +22,13 @@ def renamer(files, pattern, replacement):
                 print ("success: {} changed to -> {}".format(pathname, new_filename))
             except:
                 print ("error: {} already exists".format(new_filename))
-                try:
-                    os.remove(pathname)
-                    print ("removed: {} because its a dupe".format(pathname))
-                except:
-                    print ("error: cannot remove {}".format(pathname))
+
+                if ('isRemove'):
+                    try:
+                        os.remove(pathname)
+                        print ("removed: {} because its a dupe".format(pathname))
+                    except:
+                        print ("error: cannot remove {}".format(pathname))
 
 if __name__ == "__main__":
-    print("running...")
-    renamer("*.jpg-large", r"^(.*)\.jpg-large$", r"\1.jpg")
-    renamer("*.unsafe", r"^(.*)\.unsafe$", r"\1")
-    # renamer("*.doc", r"^(.*)\.doc$", r"new(\1).doc") #forward
-    # renamer("*.doc", r"^new\((.*)\)\.doc", r"\1.doc") #reverse
-    print("complete...")
+    main()
